Moved code for registering morphablegraphs components
